Major refactor. Separated functions into files for readability
