+++ services/codellama/src/config.py
@@ -1,0 +1,3 @@
+model_id = "codellama/CodeLlama-70b-Instruct-hf"
+cache_dir_model = "/usr/src/app/model"
+cache_dir_tokenizer = "/usr/src/app/tokenizer"
